+++ app/(afterLogin)/createPost/_components/overlay/uploadOverlay.tsx
@@ -21,32 +21,34 @@ const UploadOverlay = ({ isOverlayOpen, setIsOverlayOpen, handleOverlaySubmit }:
     }
 
     return (
-        <Overlay
-            isOpen={isOverlayOpen}
-            handleOverlaySubmit={handleOverlaySubmit}
-            onClick={() => setIsOverlayOpen(false)}
-            onLeftClick={() => setIsPreviewVisible(false)}
-            leftText={isPreviewVisible ? "다시 선택" : ""}
-            text={isPreviewVisible ? "확인했어요!" : ""}
-            textColor={"text-SYSTEM-white"}
-            imageUrl={isPreviewVisible && checkIcon}
-            leftImageUrl={isPreviewVisible && BackIcon}
-            rounded={"rounded-3xl"}
-        >
-            {!isPreviewVisible ? (
-                <ThumbnailUploader onComplete={handleComplete} />
-            ) : (
-                <>
-                    {selectedContinent && (
-                        <PreviewPostCard
-                            selectedImage={selectedImage}
-                            selectedContinent={selectedContinent}
-                            title={formData.title}
-                        />
-                    )}
-                </>
-            )}
-        </Overlay>
+        <>
+            <Overlay
+                isOpen={isOverlayOpen}
+                handleOverlaySubmit={handleOverlaySubmit}
+                onClick={() => setIsOverlayOpen(false)}
+                onLeftClick={() => setIsPreviewVisible(false)}
+                leftText={isPreviewVisible ? "다시 선택" : ""}
+                text={isPreviewVisible ? "확인했어요!" : ""}
+                textColor={"text-SYSTEM-white"}
+                imageUrl={isPreviewVisible && checkIcon}
+                leftImageUrl={isPreviewVisible && BackIcon}
+                rounded={"rounded-3xl"}
+            >
+                {!isPreviewVisible ? (
+                    <ThumbnailUploader onComplete={handleComplete} />
+                ) : (
+                    <>
+                        {selectedContinent && (
+                            <PreviewPostCard
+                                selectedImage={selectedImage}
+                                selectedContinent={selectedContinent}
+                                title={formData.title}
+                            />
+                        )}
+                    </>
+                )}
+            </Overlay>
+        </>
     )
 }
 export default UploadOverlay
