+++ components/commons/overlay.tsx
@@ -50,7 +50,7 @@ const Overlay = ({
 
     return (
         <div
-            className="font-pretendard fixed inset-0 flex justify-center items-center z-50 bg-black bg-opacity-30"
+            className="font-pretendard fixed inset-0 flex justify-center items-center z-50 bg-black bg-opacity-70"
             onClick={onClick}
             aria-modal="true"
             role="dialog"
